pulsegen/backend/pyproject.toml
[project]
name = "pulsegen"
version = "2.0.0"
description = "PulseGen — standalone ingestion microservice with Celery worker and FastAPI admin API"
requires-python = ">=3.13"
dependencies = [
    # Async HTTP
    "httpx>=0.27",
    "feedparser>=6.0",
    # Celery + Redis
    "celery[gevent]>=5.4",
    "redis>=5.0",
    # Pydantic + Settings
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    # Ollama via OpenAI-compatible API
    "openai>=1.0",
    # Source connectors
    "arxiv>=2.1",
    "ddgs>=6.0",
    # Utilities
    "python-dotenv>=1.0",
    "python-dateutil>=2.9",
    "tenacity>=9.0",
    # FastAPI + Admin API
    "fastapi>=0.135.1",
    "uvicorn>=0.42.0",
    # Database (PostgreSQL + Vector)
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.3.0",
    "sqlalchemy>=2.0.0",
]

[project.scripts]
mcp-storage-server = "mcp_servers.storage_server:main"
mcp-sql-server = "mcp_servers.sql_server:main"
mcp-search-server = "mcp_servers.search_server:main"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "ruff>=0.4",
    "mypy>=1.10",
    "freezegun>=1.5",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
# Celery's @app.task/@worker_ready.connect decorators are untyped.
module = ["src.tasks", "src.celery_app"]
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
# Test files use dynamic mocks and fixtures; strict checking adds no value.
module = "tests.*"
ignore_errors = true

[[tool.mypy.overrides]]
# Connector files use untyped third-party libraries; relax strict generics.
module = ["src.connectors.*"]
disallow_any_generics = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
    # asyncio.run is patched in pipeline task tests so coroutines are never awaited — benign
    "ignore:coroutine.*was never awaited:RuntimeWarning",
]
